requirements.txt
# Graph Database
neo4j==5.14.0
py2neo==2021.2.4

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
faker==20.1.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Visualization
plotly==5.18.0
networkx==3.2.1

# Data Validation
pydantic>=2.9.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
